requirements.txt
crowdstrike-falconpy
# python3   -m pip install SomePackage

configparser~=7.0.0

# Networks and ssh
netmiko~=4.2

# AD
ldap3~=2.9.1

# Web

requests==2.32.4
selenium==4.38.0
beautifulsoup4==4.12.2
urllib3~=2.5.0


#Openstack
aodhclient
python-novaclient>=18.3
python-cinderclient>=9.5.0
python-keystoneclient>=5.1.0
python-glanceclient>=4.4.0
python-neutronclient>=11.0.0
python-cloudkittyclient
python-designateclient
python-heatclient
python-magnumclient
python-muranoclient
python-octaviaclient
python-openstackclient>=6.2.0
python-troveclient

gnocchiclient
# python-quantumclient
python-swiftclient>=4.3.0
python-ceilometerclient>=2.9.0

# Openstack Nectar
nectar-osc
nectarallocationclient>=1.9.0
manukaclient
osc-placement>=4.1.0

# Maths
pandas>=2.0.3
numpy>=1.24.4
lxml>=4.9.3
yfinance>=0.2.28

keystoneauth1>=5.2.1
# argparse==1.4.0

scipy>=1.10.1
matplotlib>=3.7.2
statsmodels>=0.14.0
#pymc3>=3.11.6

# Network, ssh
paramiko~=3.4.0
paramiko-expect>=0.3.5

# MAAS
python-libmaas>=0.6.8

# Infoblox
infoblox-client

fqdn~=1.5.1
openstacksdk>=4.0.0
fonttools~=4.55.3
oauthlib~=3.2.2
ipaddress~=1.0.23
pyTenable~=1.6.0
prettytable~=3.11.0